DataSourceMock and HarvesterMock moved to top level class
