+++ publisher-service/src/test/java/nl/idgis/publisher/loader/messages/GetColumns.java
@@ -1,0 +1,5 @@
+package nl.idgis.publisher.loader.messages;
+
+public class GetColumns {
+	
+}
